Reset ObjectCustomField sort order when re-enabling a Custom Field
When disabling and re-enabling custom fields, the sort order stored
in the ObjectCustomFields table is retained. If the sort order in a list
of CFs is modified with MoveUp and MoveDown and a disabled CF is
then re-enabled, the previous sort order can be the same as one of
the moved OCFs.

On re-enabling a custom field, move it to the bottom of the sort
order for all ObjectCustomField records to avoid duplicate sort
order values.